requirements.txt
#pandas
#openpyxl 
#docxtpl 
#docxcompose
#docx2pdf
#python-docx 
#mailmerge 
#openpyxl 
#fastapi 
#uvicorn

flask 
docxtpl 
docx2pdf  
pandas 
openpyxl 
python-docx 
docxcompose
pypandoc


#common/docx2pdf
#common/errors
#common/files
config

merge2pdf

#fastapi
#uvicorn[standard]
#python-docx
#mailmerge
#openpyxl
#python-magic
#msal
#requests
#office365-rest-python-client
#python-dotenv
#pydantic
#Jinja2
#
